fix(firmware): guard manifest reassembly against 32-bit offset overflow
The DEPLOY_PART_MANIFEST path summed attacker-supplied offset + chunk_len
in size_t; on the 32-bit target a crafted offset near UINT32_MAX wraps the
sum past the bound and slips a wild memcpy through, before signature
verification. Promote the add to uint64_t — the same guard the model path
already uses in deploy.c. Also refresh the host-test count in the README.

+++ firmware/src/device_main.c
@@ -225,7 +225,9 @@ static void deploy_on_model_chunk(const deploy_hdr_t *h, const uint8_t *payload)
 static void deploy_on_frame(const deploy_hdr_t *h, const uint8_t *payload) {
     switch (h->part) {
         case DEPLOY_PART_MANIFEST:
-            if ((size_t)h->offset + h->chunk_len <= sizeof(g_dep.manifest)) {
+            // uint64 add so a crafted offset near UINT32_MAX can't wrap a 32-bit sum past the
+            // bound and slip a wild memcpy through — same guard the model path uses (deploy.c).
+            if ((uint64_t)h->offset + h->chunk_len <= sizeof(g_dep.manifest)) {
                 memcpy(g_dep.manifest + h->offset, payload, h->chunk_len);
                 if (h->flags & DEPLOY_FLAG_LAST) {
                     g_dep.manifest_len = h->offset + h->chunk_len;
